NumberOfAttempts(3)
increased the size of the box in the guessGame.css.

Then created a variable for attempts with the value of 3. then added code to if statement in order for the decrease of 1 if the wrong answer is guessed for either too high or too low of a guess but also a null value if guessed correct.