addedImageAndAddedCSS
added the image tag and then src in order to add the link to the image. then added width and height tag in order to make change to image because it doesn't apply with CSS changes to it. added display:flex and then changed the flex direction: row in order to put the photo on the left and the text on the right.